Parameter streamlining, bug fixes, and SWIG wrapper support. * Replaced schedule_scale_factor with anneal_cycle and T_E_inv_point. T_e_inv_point defines the point in the annealing schedule at which the annealing temperature falls below 1/e * T_0. Also updated some outdated descriptions in simanneal.physeng. * Fixed bug where random number generator for thread 0 is always improperly seeded. Removed some unnecessary RNG related code. Added boost_random dependency. * Fixed an error where v_freeze_step is not initialized * Improved suggested result output for benchmarks * Optimized and streamlined post-simulation metastability check * Added flag to swig wrapper for choosing whether to tidy results or not * Removed some Boost dependencies that are no longer required * Changed FPType to double

+++ src/global.h
@@ -16,7 +16,7 @@
 
 #include "logger.h"
 
-typedef float FPType;
+typedef double FPType;
 
 namespace saglobal {
